josh - relocate sync service under direct query sync dir
