+++ src/plugins/dashboard/public/application/utils/direct_query_sync/direct_query_sync_services.test.ts
@@ -4,11 +4,8 @@
  */
 
 import { DirectQuerySyncService } from './direct_query_sync_services';
-import {
-  extractIndexInfoFromDashboard,
-  generateRefreshQuery,
-} from '../../utils/direct_query_sync/direct_query_sync';
-import { isDirectQuerySyncEnabledByUrl } from '../../utils/direct_query_sync/direct_query_sync_url_flag';
+import { extractIndexInfoFromDashboard, generateRefreshQuery } from './direct_query_sync';
+import { isDirectQuerySyncEnabledByUrl } from './direct_query_sync_url_flag';
 
 // Mock dependencies
 jest.mock('../../utils/direct_query_sync/direct_query_sync');
